+++ pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+addopts = --ignore=tests/group_files
