+++ tests/group_files/group_bugs_in_each_step.py
