+++ tests/group_files/group_lutz_squad.py
