+++ tests/group_files/group_moon_cat.py
