+++ tests/group_files/group_no_pain_no_gain.py
