+++ tests/group_files/group_optimists_of_rationality_middle_devs.py
