+++ tests/test_group_100000/pages/dashboard_page.py
@@ -2,7 +2,7 @@ from selenium.webdriver.common.by import By
 
 from pages.base_page import BasePage
 from tests.test_group_100000.locators.dashboard_page_locators import DashboardLocators as D
-from tests.test_group_100_000 import weather_symbol
+from tests.group_files.group_100_000 import weather_symbol
 
 BTN_COOKIES = (By.CLASS_NAME, "stick-footer-panel__link")
